feat(FPD-50): create login page
