+++ .gitignore
@@ -13,7 +13,8 @@ dist-ssr
 *.local
 
 # Editor directories and files
-
+.vscode/*
+!.vscode/extensions.json
 .idea
 .DS_Store
 *.suo
