--- src/components/login/login.scss
@@ -1,13 +1,0 @@
-.actions-extra {
-  display: flex;
-  position: absolute;
-  top: 5%;
-  right: 5%;
-  justify-content: center;
-  align-items: center;
-  > * {
-    margin-left: 20px;
-    display: flex;
-    align-items: center;
-  }
-}
