+++ src/config/axios.ts
@@ -13,7 +13,6 @@ const BASE_URL = import.meta.env.VITE_API_URL;
 
 const instance = axios.create({
   baseURL: `${BASE_URL}/`,
-  withCredentials: true,
   timeout: 300000,
   headers: {
     "Content-Type": "application/json",
