+++ src/pages/login/login.scss
@@ -1,0 +1,36 @@
+.login {
+  display: flex;
+  background-color: transparent;
+}
+.actions-extra {
+  display: flex;
+  position: absolute;
+  top: 5%;
+  right: 5%;
+  justify-content: center;
+  align-items: center;
+  > * {
+    margin-left: 20px;
+    display: flex;
+    align-items: center;
+  }
+}
+.canva_core {
+  position: absolute;
+  top: 0px;
+  left: 0px;
+  transform-origin: 0px 0px;
+  height: 538.216px;
+  width: 549.622px;
+  transform: scale(0.74817);
+}
+.canva {
+  position: absolute;
+}
+.ant-input-affix-wrapper-lg {
+  padding: 15px 11px;
+}
+.ant-btn-lg {
+  height: 50px;
+  font-size: 20px;
+}
