+++ src/redux/features/auth/authSlice.ts
@@ -1,10 +1,10 @@
 import { PayloadAction, createSlice } from "@reduxjs/toolkit";
-import { getGlobalState } from "../../../utils/getGloabal";
 import { jwtDecode } from "jwt-decode";
+import { getGlobalState } from "../../../utils/getGloabal";
 
 interface User {
   id: string;
-  name: string;
+  username: string;
   email: string;
   role: string;
 }
@@ -19,7 +19,7 @@ interface AuthState {
 }
 
 const initialState: AuthState = {
-  isAuth: localStorage.getItem("token") ? true : false,
+  isAuth: localStorage.getItem("accessToken") ? true : false,
   ...getGlobalState(),
   user: null,
   role: "",
@@ -34,8 +34,9 @@ const authSlice = createSlice({
       const data = jwtDecode<{ username: string; id: string; email: string; role: string }>(
         _action.payload.access_token,
       );
-      state.user = { name: data.username, id: data.id, email: data.email, role: data.role };
+      state.user = { username: data.username, id: data.id, email: data.email, role: data.role };
       state.isAuth = true;
+      state.role = data.role;
     },
     setUserItem(state, action: PayloadAction<Partial<AuthState>>) {
       Object.assign(state, action.payload);
